Don't explicitly encode healthchecks.
The latest docker containers for postgres and redis encode the
healthcheck in the image itself, so we don't need to repeat it in the
docker compose files.

Note that this only works if you are using the latest docker images. But
at worst you'll just have to restart the API service after starting up
postgres/redis.